+++ internal/bullet/bullet.go
@@ -2,5 +2,5 @@ package bullet
 
 import "unsafe"
 
-func MallocTrampoline(fn unsafe.Pointer, n uint64) unsafe.Pointer
+func MallocTrampoline(fn unsafe.Pointer, n uint) unsafe.Pointer
 func FreeTrampoline(fn unsafe.Pointer, ptr unsafe.Pointer)
